fix callback functions deregistering tasks
previously callbacks created a new Python instance of the corresponding
task. This caused two (or more) Python classes for the same nidaq task
to exist. As soon as one of them went out of scope the cleanup code was
called and the task was cleared, causing all later calls using the
existing instance to fail.

This fixes the issue by maintaining a cache off all registered tasks. A
small wrapper function is generated. This function picks the correct Task
instance based on the task id of the task generating the callback.